requirements.txt
langchain==0.1.17
langchain-community==0.0.36

faiss-cpu==1.8.0
tiktoken==0.7.0
python-dotenv==1.0.1

openai==0.28.1
pydantic==1.10.15
